Remove module.
