+++ src/dcegm/pre_processing.py
@@ -5,7 +5,6 @@ from typing import Tuple
 
 import numpy as np
 import pandas as pd
-from dcegm.aggregate_policy_value import calc_current_value
 
 
 def get_partial_functions(
@@ -80,6 +79,37 @@ def get_partial_functions(
         compute_next_wealth_matrices,
         transition_function,
     )
+
+
+def calc_current_value(
+    consumption: np.ndarray,
+    next_period_value: np.ndarray,
+    choice: int,
+    discount_factor: float,
+    compute_utility: Callable,
+) -> np.ndarray:
+    """Compute the agent's current value.
+
+    We only support the standard value function, where the current utility and
+    the discounted next period value have a sum format.
+
+    Args:
+        consumption (np.ndarray): Level of the agent's consumption.
+            Array of shape (n_quad_stochastic * n_grid_wealth,).
+        next_period_value (np.ndarray): The value in the next period.
+        choice (int): The current discrete choice.
+        compute_utility (callable): User-defined function to compute the agent's
+            utility. The input ``params``` is already partialled in.
+        discount_factor (float): The discount factor.
+
+    Returns:
+        - value (np.ndarray): The current value.
+
+    """
+    utility = compute_utility(consumption, choice)
+    value = utility + discount_factor * next_period_value
+
+    return value
 
 
 def create_multi_dim_arrays(
