+++ tests/test_upper_envelope.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 import pytest
-from dcegm.aggregate_policy_value import calc_current_value
+from dcegm.pre_processing import calc_current_value
 from dcegm.upper_envelope import upper_envelope
 from numpy.testing import assert_array_almost_equal as aaae
 from toy_models.consumption_retirement_model.utility_functions import utility_func_crra
